+++ package.json
@@ -13,6 +13,7 @@
     "cookie-parser": "^1.4.4",
     "cors": "^2.8.5",
     "dataloader": "^1.4.0",
+    "date-fns": "^2.12.0",
     "express": "^4.17.1",
     "express-flash": "^0.0.2",
     "express-oas-generator": "^1.0.9",
