+++ src/app.js
@@ -27,7 +27,6 @@ import corsOptions from './cors';
 import user from './routes/user'; // TODO: REMOVE the user route.
 import tokenRequest from './routes/tokenRequest';
 import healthcheck from './routes/healthCheck';
-import * as UserService from './services/user-service';
 
 i18next
   .use(LanguageDetector)
@@ -88,7 +87,6 @@ app.use('/user', user);
 app.use('/token-request', tokenRequest);
 app.use('/healthcheck', healthcheck);
 
-UserService.prepareHourDayWeakLimitData();
 // The following routes are intended to be used in development mode only
 // TODO : Remove Me
 if (process.env.NODE_ENV !== 'production') {
