+++ src/constants/error-status.js
@@ -10,3 +10,4 @@ export const OVERALL_LIMIT_REACHED = 'OVERALL_LIMIT_REACHED';
 export const HOURLY_LIMIT_REACHED = 'HOURLY_LIMIT_REACHED';
 export const DAILY_LIMIT_REACHED = 'DAILY_LIMIT_REACHED';
 export const WEEKLY_LIMIT_REACHED = 'WEEKLY_LIMIT_REACHED';
+export const HOUR_DAY_WEEK_DATA_NOT_PREPARED = 'HOUR_DAY_WEEK_DATA_NOT_PREPARED';
