+++ src/services/user-service.js
@@ -1,8 +1,6 @@
-import { BN } from 'bn.js';
 import { Users, TokenRequests } from '../models';
 import Config from '../config/config';
-import * as ErrorStatus from '../constants/error-status';
-let hourDayWeekData = { daily: {}, hourly: {}, weekly: {}};
+import { updateHourDayWeakData } from './token-limit-service'; 
 
 export const getCentrifugeUser = async (githubUser) => {
   try{
@@ -78,116 +76,5 @@ export const logTokenRequest = async (centrifugeUser, tokenDetails) => {
     }
     catch(ex) {
       throw new Error('INVALID_TOKEN_REQUEST');
-    }
-};
-
-export const checkOverallTokenLimit = async (centrifugeUser) => {
-  try{
-    const accountTokenLimit = Config.CFG_ACCOUNT_TOKEN_LIMIT;
-    const TokenRequest = await TokenRequests.findAll({
-        where : { user_uuid: centrifugeUser.centrifugeUser.uuid },
-        raw : true 
-        });
-      let totalValue;
-      TokenRequest.map(req => {
-        totalValue = new BN(totalValue).add(new BN(req.txAmount));
-      });
-      if(totalValue.gt(new BN(accountTokenLimit))) 
-        return false;
-      return true;
-    }
-    catch(ex) {
-      throw new Error(ErrorStatus.OVERALL_LIMIT_REACHED);
-    }
-};
-
-const getDateHour = (createdAt) => {
-  const day = new Date(createdAt);
-  return day.getFullYear() + '-' + (day.getMonth() + 1) + '-' + day.getDate() + ' ' + day.getHours() + ':00:00';
-};
-
-const getDay = (createdAt) => {
-  const day = new Date(createdAt);
-  return day.getFullYear() + '-' + (day.getMonth() + 1) + '-' + day.getDate() + ' 00:00:00';
-};
-
-function getWeekNumber(d) {
-  // Copy date so don't modify original
-  d = new Date(Date.UTC(d.getFullYear(), d.getMonth(), d.getDate()));
-  // Set to nearest Thursday: current date + 4 - current day number
-  // Make Sunday's day number 7
-  d.setUTCDate(d.getUTCDate() + 4 - (d.getUTCDay()||7));
-  // Get first day of year
-  var yearStart = new Date(Date.UTC(d.getUTCFullYear(),0,1));
-  // Calculate full weeks to nearest Thursday
-  var weekNo = Math.ceil(( ( (d - yearStart) / 86400000) + 1)/7);
-  // Return array of year and week number
-  return [d.getUTCFullYear(), weekNo];
-}
-
-const getWeekNo = (createdAt) => {
-  const day = new Date(createdAt);
-  var result = getWeekNumber(day);
-  return result[1] + '-' + result[0];
-};
-
-export const updateHourDayWeakData = (createdAt) => {
-  const dayhour = getDateHour(createdAt);
-  const transferAmount = Config.CFG_TRANSFER_AMOUNT;
-  if(hourDayWeekData.hourly[dayhour] === undefined) {
-    hourDayWeekData.hourly[dayhour] = new BN(transferAmount);
-  }
-  else {
-    hourDayWeekData.hourly[dayhour] = new BN(hourDayWeekData.hourly[dayhour]).add(new BN(transferAmount));
-  }
-
-  const day = getDay(createdAt);
-  if(hourDayWeekData.daily[day] === undefined) {
-    hourDayWeekData.daily[day] = new BN(transferAmount);
-  }
-  else {
-    hourDayWeekData.daily[day] = new BN(hourDayWeekData.daily[day]).add(new BN(transferAmount));
-  }
-
-  const weekNo = getWeekNo(createdAt);
-  if(hourDayWeekData.weekly[weekNo] === undefined) {
-    hourDayWeekData.weekly[weekNo] = new BN(transferAmount);
-  }
-  else {
-    hourDayWeekData.weekly[weekNo] = new BN(hourDayWeekData.weekly[weekNo]).add(new BN(transferAmount));
-  }
-};
-
-export const checkHourDayWeakLimit = async () => {
-  const hourlyLimit = Config.CFG_HOURLY_LIMIT;
-  const dailyLimit = Config.CFG_DAILY_LIMIT;
-  const weeklyLimit = Config.CFG_WEEKLY_LIMIT;
-  const reqDate = new Date();
-  const dayHour = getDateHour(reqDate);
-  if(new BN(hourDayWeekData.hourly[dayHour]).gte(new BN(hourlyLimit))) {
-    throw new Error(ErrorStatus.HOURLY_LIMIT_REACHED);
-  }
-  const day = getDay(reqDate);
-  if(new BN(hourDayWeekData.daily[day]).gte(new BN(dailyLimit))) {
-    throw new Error(ErrorStatus.DAILY_LIMIT_REACHED);
-  }
-  const weekNo = getWeekNo(reqDate);
-  if(new BN(hourDayWeekData.weekly[weekNo]).gte(new BN(weeklyLimit))) {
-    throw new Error(ErrorStatus.WEEKLY_LIMIT_REACHED);
-  }
-  return true;
-};
-
-export const prepareHourDayWeakLimitData = async () => {
-  try{
-      const TokenRequest = await TokenRequests.findAll({
-        raw : true 
-      });
-      TokenRequest.map(req=> {
-        updateHourDayWeakData(req.createdAt);
-      });
-    }
-    catch(ex) {
-      throw new Error(ErrorStatus.OVERALL_LIMIT_REACHED);
     }
 };
